Replace Arrow file format with Arrow stream format instead of having a new conf.
