+++ core/src/main/scala/org/apache/spark/api/python/PythonRunner.scala
@@ -36,7 +36,6 @@ private[spark] object PythonEvalType {
   val NON_UDF = 0
   val SQL_BATCHED_UDF = 1
   val SQL_PANDAS_UDF = 2
-  val SQL_PANDAS_UDF_STREAM = 3
 }
 
 /**
