+++ sql/catalyst/src/main/scala/org/apache/spark/sql/internal/SQLConf.scala
@@ -925,13 +925,6 @@ object SQLConf {
       .intConf
       .createWithDefault(10000)
 
-  val ARROW_EXECUTION_STREAM_ENABLE =
-    buildConf("spark.sql.execution.arrow.stream.enable")
-      .internal()
-      .doc("When using Apache Arrow, use Arrow stream protocol if possible.")
-      .booleanConf
-      .createWithDefault(false)
-
   object Deprecated {
     val MAPRED_REDUCE_TASKS = "mapred.reduce.tasks"
   }
@@ -1209,8 +1202,6 @@ class SQLConf extends Serializable with Logging {
   def arrowEnable: Boolean = getConf(ARROW_EXECUTION_ENABLE)
 
   def arrowMaxRecordsPerBatch: Int = getConf(ARROW_EXECUTION_MAX_RECORDS_PER_BATCH)
-
-  def arrowStreamEnable: Boolean = getConf(ARROW_EXECUTION_STREAM_ENABLE)
 
   /** ********************** SQLConf functionality methods ************ */
